src/gcp-data-chat/app/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0
google-cloud-secret-manager>=2.16.0
google-auth>=2.22.0
db-dtypes>=1.1.0
openai>=1.0.0
